pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "persona-agent"
version = "0.1.0"
description = "A local role-playing AI agent with dynamic persona switching, mood management, and linguistic style customization"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "agent", "roleplay", "persona", "llm", "claude", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "httpx>=0.25",
    "aiofiles>=23.0",
    "aiosqlite>=0.19",
    "chromadb>=0.4",
    "tiktoken>=0.5",
    "watchdog>=3.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
persona-agent = "persona_agent.ui.cli:main"
pa = "persona_agent.ui.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/persona-agent"
Repository = "https://github.com/yourusername/persona-agent"
Documentation = "https://github.com/yourusername/persona-agent/blob/main/docs"
Issues = "https://github.com/yourusername/persona-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/persona_agent"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501", "SIM102"]  # Line too long (handled by black), complex ifs

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/persona_agent"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
